fix yang out of range

+++ database/migrations/2018_10_24_023811_create_trx_work_order_table.php
@@ -21,7 +21,7 @@ class CreateTrxWorkOrderTable extends Migration
             $table->unsignedInteger('project_id')->nullable();
             $table->text('description')->nullable();
             $table->integer('status')->default(1);
-            $table->bigInteger('total_price')->nullable();
+            $table->double('total_price')->nullable();
             $table->unsignedInteger('branch_id');
             $table->unsignedInteger('user_id');
             $table->timestamps();
